% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{etrd_alerts}
\alias{etrd_alerts}
\title{ETrade Alert Search}
\usage{
etrd_alerts(
  account,
  count = 25,
  category = NULL,
  status = NULL,
  search = NULL,
  output = c("df", "list"),
  access_tokens = NULL,
  etrade_cred = NULL,
  sandbox = FALSE
)
}
\arguments{
\item{account}{The account ID key, not the account ID associated with the
specific account. Use \code{\link{etrd_account_list}} to identify the
proper account id key.}

\item{count}{The alert count. By default it returns 25. Max values that can
be returned: 300.}

\item{category}{The alert category. By default it will return STOCK and
ACCOUNT. Options: STOCK, ACCOUNT}

\item{status}{The alert status. By default it will return READ and UNREAD.
Options: READ, UNREAD, DELETED}

\item{search}{The alert search. Search is done based on the subject.}

\item{output}{Indicate whether the output should be in the form of a data
frame ('df') or list ('list'). Data frame is returned by default.}

\item{access_tokens}{Access tokens are created using
\code{\link{etrd_auth_access_token}}. This entry is not required because
the output is saved and retrieved from R options automatically.}

\item{etrade_cred}{The output created from
\code{\link{etrd_auth_credentials}} when a valid ETRADE key and secret have
been passed. This entry is not required because the output is saved and
retrieved from R options automatically.}

\item{sandbox}{ETRADE offers a sandbox environment for validating API calls
and responses. If using the sandbox environment, this must be set to TRUE
in each function called throughout etrader. ETRADE states "Sandbox
responses use stored data that's intended to provide typical responses for
basic use cases. So the responses you receive will not contain current
data, and may not exactly match your requests in other ways." Essentially,
the responses will not match the requests entered but successful pull will
indicate whether the entry was valid or not.}
}
\value{
a list or data frame of alert data
}
\description{
Search alerts associated with the provided ETrade Account
}
\examples{
\dontrun{

account_id = etrd_account_list()$accountIdKey[1]
etrd_alerts(account_id)

}
}
